♻️ (crypto): remove: typecase * as u32 -> use *u32 directly

+++ crypto/src/elgamal/encryption.rs
@@ -81,7 +81,7 @@ impl ElGamal {
     /// * `g` - The generator of the cyclic group Z_p (BigUint)
     /// * `p` - The group modulus p (BigUint)
     fn decode_message(encoded_message: &BigUint, g: &BigUint, p: &BigUint) -> BigUint {
-        let one = 1 as u32;
+        let one = 1u32;
         let mut message = BigUint::zero();
 
         // *encoded_message = dereference 'encoded_message' to get the value
@@ -103,21 +103,21 @@ mod tests {
     #[test]
     fn it_should_encode_a_message() {
         let params = ElGamalParams {
-            p: BigUint::from(7 as u32),
+            p: BigUint::from(7u32),
             // and, therefore, q -> 3
-            g: BigUint::from(2 as u32),
+            g: BigUint::from(2u32),
         };
-        let message = BigUint::from(2 as u32);
+        let message = BigUint::from(2u32);
         let encoded_message = ElGamal::encode_message(&message, &params.g, &params.p);
-        assert_eq!(encoded_message, BigUint::from(4 as u32));
+        assert_eq!(encoded_message, BigUint::from(4u32));
     }
 
     #[test]
     fn it_should_decode_0() {
         let params = ElGamalParams {
-            p: BigUint::from(7 as u32),
+            p: BigUint::from(7u32),
             // and, therefore, q -> 3
-            g: BigUint::from(2 as u32),
+            g: BigUint::from(2u32),
         };
         let zero = BigUint::zero();
         let message = zero.clone();
@@ -129,9 +129,9 @@ mod tests {
     #[test]
     fn it_should_decode_1() {
         let params = ElGamalParams {
-            p: BigUint::from(7 as u32),
+            p: BigUint::from(7u32),
             // and, therefore, q -> 3
-            g: BigUint::from(2 as u32),
+            g: BigUint::from(2u32),
         };
         let one = BigUint::one();
         let message = one.clone();
@@ -143,13 +143,13 @@ mod tests {
     #[test]
     fn it_should_decode_25() {
         let params = ElGamalParams {
-            p: BigUint::from(23 as u32),
+            p: BigUint::from(23u32),
             // and, therefore, q -> 11
-            g: BigUint::from(2 as u32),
+            g: BigUint::from(2u32),
         };
 
         // choose a message m > 1 && m < q
-        let nine = BigUint::from(9 as u32);
+        let nine = BigUint::from(9u32);
         let message = nine.clone();
         let encoded_message = ElGamal::encode_message(&message, &params.g, &params.p);
         let decoded_message = ElGamal::decode_message(&encoded_message, &params.g, &params.p);
@@ -159,50 +159,50 @@ mod tests {
     #[test]
     fn it_should_encrypt() {
         let params = ElGamalParams {
-            p: BigUint::from(7 as u32),
+            p: BigUint::from(7u32),
             // and, therefore, q -> 3
-            g: BigUint::from(2 as u32),
+            g: BigUint::from(2u32),
         };
 
         // generate a public/private key pair
-        let r = BigUint::from(2 as u32);
+        let r = BigUint::from(2u32);
         let (pk, _sk) = Helper::generate_key_pair(&params, &r);
 
         // the value of the message: 1
-        let message = BigUint::from(1 as u32);
+        let message = BigUint::from(1u32);
 
         // a new random value for the encryption
-        let r_ = BigUint::from(1 as u32);
+        let r_ = BigUint::from(1u32);
 
         // encrypt the message
         let encrypted_message = ElGamal::encrypt(&message, &r_, &pk);
 
         // check that a = g^r_ = 2^1 mod 7 = 2
-        assert_eq!(encrypted_message.a, BigUint::from(2 as u32));
+        assert_eq!(encrypted_message.a, BigUint::from(2u32));
 
         // check that b = h^r_*g^m = (g^r)^r_ * g^m
         // b = ((2^2)^1 mod 7 * 2^1 mod 7) mod 7
         // b = (4 mod 7 * 2 mod 7) mod 7 = 1
-        assert_eq!(encrypted_message.b, BigUint::from(1 as u32));
+        assert_eq!(encrypted_message.b, BigUint::from(1u32));
     }
 
     #[test]
     fn it_should_encrypt_decrypt_2() {
         let params = ElGamalParams {
-            p: BigUint::from(23 as u32),
+            p: BigUint::from(23u32),
             // and, therefore, q -> 11
-            g: BigUint::from(2 as u32),
+            g: BigUint::from(2u32),
         };
 
         // generate a public/private key pair
-        let r = BigUint::from(9 as u32);
+        let r = BigUint::from(9u32);
         let (pk, sk) = Helper::generate_key_pair(&params, &r);
 
         // the value of the message: 2
-        let message = BigUint::from(2 as u32);
+        let message = BigUint::from(2u32);
 
         // a new random value for the encryption
-        let r_ = BigUint::from(5 as u32);
+        let r_ = BigUint::from(5u32);
 
         // encrypt the message
         let encrypted_message = ElGamal::encrypt(&message, &r_, &pk);
